moved parallel match prediction algorithm to new class
